Refactor text-provider into text-preprocessor pipeline
Using pipeline and data in the same fashion as other pipeline steps.
