+++ pom.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
@@ -19,7 +20,8 @@
     inconsistencies. ArDoCo is actively developed by researchers of the Modelling for Continuous Software Engineering
     (MCSE) group of KASTEL - Institute of Information Security and Dependability at the KIT. This work was supported by
     funding from the topic Engineering Secure Systems of the Helmholtz Association (HGF) and by KASTEL Security Research
-    Labs (46.23.01).</description>
+    Labs (46.23.01).
+  </description>
   <licenses>
     <license>
       <name>MIT License</name>
@@ -94,7 +96,7 @@
       </dependency>
       <dependency>
         <groupId>${project.groupId}</groupId>
-        <artifactId>text-provider</artifactId>
+        <artifactId>text-preprocessing</artifactId>
         <version>${project.version}</version>
       </dependency>
       <dependency>
@@ -224,7 +226,7 @@
       </activation>
       <modules>
         <module>text-extraction</module>
-        <module>text-provider</module>
+        <module>text-preprocessing</module>
         <module>recommendation-generator</module>
         <module>common</module>
         <module>pipeline</module>
@@ -260,7 +262,7 @@
       </activation>
       <modules>
         <module>text-extraction</module>
-        <module>text-provider</module>
+        <module>text-preprocessing</module>
         <module>recommendation-generator</module>
         <module>common</module>
         <module>pipeline</module>
