+++ text-preprocessing/src/main/java/edu/kit/kastel/mcse/ardoco/core/text/providers/TextPreprocessingAgent.java
@@ -1,0 +1,38 @@
+package edu.kit.kastel.mcse.ardoco.core.text.providers;
+
+import java.util.List;
+import java.util.Map;
+
+import edu.kit.kastel.informalin.data.DataRepository;
+import edu.kit.kastel.informalin.framework.configuration.Configurable;
+import edu.kit.kastel.informalin.pipeline.AbstractPipelineStep;
+import edu.kit.kastel.mcse.ardoco.core.api.agent.Informant;
+import edu.kit.kastel.mcse.ardoco.core.api.agent.PipelineAgent;
+import edu.kit.kastel.mcse.ardoco.core.text.providers.informants.corenlp.CoreNLPProvider;
+
+public class TextPreprocessingAgent extends PipelineAgent {
+
+    private final List<Informant> informants;
+
+    @Configurable
+    private List<String> enabledInformants;
+
+    /**
+     * Instantiates a new initial text agent.
+     */
+    public TextPreprocessingAgent(DataRepository data) {
+        super(TextPreprocessingAgent.class.getSimpleName(), data);
+        informants = List.of(new CoreNLPProvider(data));
+        enabledInformants = informants.stream().map(AbstractPipelineStep::getId).toList();
+    }
+
+    @Override
+    protected void delegateApplyConfigurationToInternalObjects(Map<String, String> additionalConfiguration) {
+        informants.forEach(e -> e.applyConfiguration(additionalConfiguration));
+    }
+
+    @Override
+    protected List<Informant> getEnabledPipelineSteps() {
+        return findByClassName(enabledInformants, informants);
+    }
+}
