+++ text-preprocessing/src/main/resources/edu.kit.kastel.mcse.ardoco.core.text.providers.indirect.agents.indirect.Stanford.conf
