+++ text-preprocessing/src/test/java/edu/kit/kastel/mcse/ardoco/core/text/providers/corenlp/CoreNLPPhraseTest.java
@@ -1,0 +1,13 @@
+/* Licensed under MIT 2022. */
+package edu.kit.kastel.mcse.ardoco.core.text.providers.corenlp;
+
+import edu.kit.kastel.mcse.ardoco.core.api.data.text.NlpInformant;
+import edu.kit.kastel.mcse.ardoco.core.text.providers.base.PhraseTest;
+
+class CoreNLPPhraseTest extends PhraseTest {
+
+    @Override
+    protected NlpInformant getProvider() {
+        return CoreNLPProviderTest.getCoreNLPProvider();
+    }
+}
