+++ text-preprocessing/src/test/java/edu/kit/kastel/mcse/ardoco/core/text/providers/corenlp/CoreNLPTextTest.java
@@ -1,0 +1,12 @@
+/* Licensed under MIT 2022. */
+package edu.kit.kastel.mcse.ardoco.core.text.providers.corenlp;
+
+import edu.kit.kastel.mcse.ardoco.core.api.data.text.NlpInformant;
+import edu.kit.kastel.mcse.ardoco.core.text.providers.base.TextTest;
+
+class CoreNLPTextTest extends TextTest {
+    @Override
+    protected NlpInformant getProvider() {
+        return CoreNLPProviderTest.getCoreNLPProvider();
+    }
+}
