--- text-provider/src/test/java/edu/kit/kastel/mcse/ardoco/core/text/providers/corenlp/CoreNLPWordTest.java
@@ -1,12 +1,0 @@
-/* Licensed under MIT 2022. */
-package edu.kit.kastel.mcse.ardoco.core.text.providers.corenlp;
-
-import edu.kit.kastel.mcse.ardoco.core.api.data.text.TextProvider;
-import edu.kit.kastel.mcse.ardoco.core.text.providers.base.WordTest;
-
-class CoreNLPWordTest extends WordTest {
-    @Override
-    protected TextProvider getProvider() {
-        return CoreNLPProviderTest.getCoreNLPProvider();
-    }
-}
